backup_20250713_235505/requirements_original.txt
torch>=1.9.0,<2.0.0
scikit-learn>=1.0.0,<1.3.0
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
networkx>=2.6.0
scipy>=1.7.0
pyyaml>=6.0

# Real API Dependencies
alpha-vantage==2.3.1
yfinance==0.2.28
quandl==3.7.0
twelvedata>=1.2.12,<1.3.0
polygon-api-client==1.12.3
finnhub-python==2.4.20
iex-cloud==0.0.4
newsapi-python==0.2.7
tweepy==4.14.0
praw==7.7.1
pycryptodome==3.19.0
requests-oauthlib==1.3.1
websocket-client==1.6.4
python-binance==1.0.19
ccxt==4.1.64
ta-lib==0.4.28
fredapi==0.5.1
pandas-datareader==0.10.0
pycoingecko==3.1.0
cryptocompare==0.7.6
sentiment-analysis==0.1.0
textblob==0.17.1
vaderSentiment==3.3.2
newspaper3k==0.2.8
beautifulsoup4==4.12.2
selenium==4.15.2
scrapy==2.11.0
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1
celery==5.3.4
python-telegram-bot==20.7
discord.py==2.3.2
slack-sdk==3.23.0
google-cloud-storage==2.10.0
boto3==1.34.0
azure-storage-blob==12.19.0
kafka-python==2.0.2
prometheus-client==0.19.0
grafana-api==1.0.3
datadog==0.49.1
newrelic==9.2.0
sentry-sdk==1.38.0

# Essential API packages (already added by main script)
aiohttp==3.9.1
python-dotenv==1.0.0
feedparser==6.0.10
requests==2.31.0

# Add these if not already present
yfinance>=0.2.18
ccxt>=4.1.0
python-binance>=1.0.17
pycoingecko>=3.1.0
textblob>=0.17.1
vaderSentiment>=3.3.2
tweepy>=4.14.0
praw>=7.7.0
newsapi-python>=0.2.6

# Basic scientific computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0
loguru>=0.7.0
